requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
scikit-learn
python-multipart
pydantic
python-dotenv
httpx
requests
joblib